backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyGithub==1.59.1
boto3==1.34.131
botocore==1.34.131
cryptography==41.0.7 